Fixing test

+++ src/main/org/jboss/messaging/core/management/MessagingServerControlMBean.java
@@ -92,7 +92,7 @@ public interface MessagingServerControlMBean
    
    int getAIOBufferSize();
    
-   long getAIOBufferTimeout();
+   int getAIOBufferTimeout();
    
    public long getPagingMaxGlobalSizeBytes();
